% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pilze_BRD.R
\name{pilze_BRD}
\alias{pilze_BRD}
\title{Abfrage Pilzartenliste Deutschlands nach MTB}
\usage{
pilze_BRD(suche)
}
\arguments{
\item{suche}{ist das MTB (MessTischBlatt) für das eine Artenliste abgefragt werden soll}
}
\value{
Gibt einen Vektor mit Arten zurueck
}
\description{
Abfrage Pilzartenliste Deutschlands nach MTB
}
\details{
hier kommen noch Details
}
\examples{
\dontrun{
pilze_BRD(suche = "Regensburg")
}
}
\references{
\url{http://brd.pilzkartierung.de}
}

